chore(tests): move root JSON-LD tests into JsonLd/ subdir
* HydraTest -> JsonLd/HydraHideFromDocsTest (descriptive name reflects what
  the single test asserts: hiding hydra:supportedClass and operations)
* LinkedDataPlatformTest -> JsonLd/LinkedDataPlatformTest (LDP is JSON-LD)
* ItemUriTemplateTest -> JsonLd/ItemUriTemplateNotFoundTest (rename
  testIssue6718 to testNotFoundOnInvalidItemUriTemplateRelation)
